Title: 특정 수 도달하기, Time: 7ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 250212/특정 수 도달하기/reaching-specific-number.cpp
@@ -18,7 +18,7 @@ int main() {
         }
         sum+=arr[i];
         if(i==9){
-            cout << sum << " " << sum/10;
+            cout << sum << " " << (double)sum/10;
             break;
         }
